pyproject.toml
[tool.poetry]
name = "displayimageadaeink"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.2.0"
board = "^1.0"
adafruit-circuitpython-epd = "^2.9.3"
flask = "^2.0.1"
werkzeug = "^2.0.1"
watchdog = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
